moved deploy functions to deploymenu
